The flowers now move and bounce
The flowers now move and bounce when they reach the borders. I added two functions to the Flower class: bounce() and move(). Besides that, I created 12 flowers in total, which are displayed depending on where you have the cursor. And I got rid of the background color in the draw function, so that it would draw every frame of the flowers, giving it a cool effect.